1907- move GetDefaultControls and GetCurrentlyAppliedControls (#1931)
* move GetDefaultControls and GetCurrentlyAppliedControls from InputGroupList class to Settings class

* added blank line before comment

* changed function calling

* changed function calling

* removed blank line
