+++ src/general/OptionsMenu.cs
@@ -796,7 +796,7 @@ public class OptionsMenu : Control
 
     private void InputDefaultsConfirm()
     {
-        Settings.Instance.CurrentControls.Value = InputGroupList.GetDefaultControls();
+        Settings.Instance.CurrentControls.Value = Settings.GetDefaultControls();
         Settings.Instance.ApplyInputSettings();
         BuildInputRebindControls();
